https://github.com/zitadel/zitadel/issues/6761
[optional config]
DisableExternalIDPPanel: false # ZITADEL_LOGIN_DISABLEEXTERNALIDPPANEL

This will GLOBALLY disable the external IDP panel.
The main use case is that there may be thousands of external idps available for a given org.
The Login UI should not show these but instead an auth request passing the idp scope can be used to cherry pick which external IDP a user is challenged with.