eature: generic oauth configurable prompt
Unfortunately, there seem to be some major oauth companies that throw an error when the authorizer is passed a "prompt" value (f any kind, seemingly). Ping is one of them. This commit does not add a select menu to the UI for this, but it does lay out the switch statement, driven by env var, that is consistent with the authorization.proto constants defined in oidc v2beta.